requirements.txt
pymongo<3.0
Flask-MongoKit==0.6
Flask-WTF==0.11
Flask-Cache==0.13.1
gunicorn
redis==2.10.3
hiredis==0.2.0
numexpr
cython
tables==3.2.0
rq==0.5.3
rq-dashboard==0.3.4
boto
geojson
paegan-transport>=1.0.1
